refactor: replace int with long for timestamps to avoid useless conversions

+++ snail-race-bet/java/src/main/java/snail/race/kata/domain/BetApplication.java
@@ -10,7 +10,7 @@ public class BetApplication {
 
     void placeBet(
              String gambler,
-             int timestamp,
+             long timestamp,
              int first,
              int second,
              int third) {
